Update hwo the parameter is handled in McpCommand.
BaseTool have a schema propertly and is populate
through a get method.
In Mcp tool handling, we override this get
function to use the parametersJsonSchema to construct
schema: FunctionDeclaration.

Getting the parameter schema from schema: FunctionDeclaration is
a more ideal way, as schema is a property of BaseTool,
parameterSchema is an implementation detail of the BaseTool. The
schema: FunctionDeclaration is more stable.